requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
asgiref==3.8.1
openai==1.66.2
selenium==4.28.1
g4f==0.3.2.9
g4f[all]
google-generativeai==0.8.3
anthropic==0.42.0
aiohttp==3.11.11
pytest==8.3.4
pytest-asyncio==0.24.0
streamlit==1.31.0
psutil==5.9.8
pandas==2.2.0